separate CMEK options struct tests from the upload/http mock tests
