+++ packages/storage/Tests/CustomerEncryptionKeyTests.swift
@@ -110,147 +110,4 @@ import Testing
     let key2 = try CustomerEncryptionKeyOptions(keyBase64: sample.keyBase64)
     #expect(key1 == key2)
   }
-
-  @Test func simpleUploadWithCSEKHeaders() async throws {
-    let registry = MockRegistry.create()
-    let bucket = "test-bucket"
-    let objectName = "test-csek-object"
-    let data = Data(repeating: 1, count: 1024)
-    let source = BytesSource(data: data)
-
-    let sample = sampleKey()
-    let csek = try CustomerEncryptionKeyOptions(key: sample.data)
-
-    let simpleUploadUrl = registry.url(
-      "/upload/storage/v1/b/\(bucket)/o?uploadType=multipart&name=\(objectName)")
-
-    let responseJSON = """
-      {
-        "name": "\(objectName)",
-        "bucket": "\(bucket)",
-        "generation": "1",
-        "metageneration": "1",
-        "size": "\(data.count)",
-        "contentType": "application/octet-stream",
-        "customerEncryption": {
-          "encryptionAlgorithm": "AES256",
-          "keySha256": "\(sample.keyHashBase64)"
-        }
-      }
-      """
-
-    registry.register(
-      response: .success(
-        statusCode: 200, data: responseJSON.data(using: .utf8)!,
-        headers: nil),
-      for: simpleUploadUrl)
-
-    let config = URLSessionConfiguration.ephemeral
-    config.protocolClasses = [MockURLProtocol.self]
-    let session = URLSession(configuration: config)
-
-    let options = StorageClientOptions().with {
-      $0.client = .init().with {
-        $0.endpoint = registry.endpoint
-        $0._testSession = session
-      }
-    }
-
-    let client = try StorageClient(options)
-    let uploadOptions = UploadOptions(customerEncryptionKey: csek)
-    let task = client.upload(source, to: bucket, as: objectName, options: uploadOptions)
-    let object = try await task.value
-
-    #expect(object.name == objectName)
-    #expect(object.customerEncryption?.encryptionAlgorithm == "AES256")
-    #expect(object.customerEncryption?.keySha256 == sample.keyHashBase64)
-
-    let requests = registry.recordedRequests()
-    #expect(!requests.isEmpty)
-    let uploadReq = requests.first {
-      $0.url?.path.contains("/upload/storage/v1/b/\(bucket)/o") == true
-    }
-    #expect(uploadReq != nil)
-    #expect(uploadReq?.value(forHTTPHeaderField: "x-goog-encryption-algorithm") == "AES256")
-    #expect(uploadReq?.value(forHTTPHeaderField: "x-goog-encryption-key") == sample.keyBase64)
-    #expect(
-      uploadReq?.value(forHTTPHeaderField: "x-goog-encryption-key-sha256") == sample.keyHashBase64)
-  }
-
-  @Test func resumableUploadWithCSEKHeaders() async throws {
-    let registry = MockRegistry.create()
-    let bucket = "test-bucket"
-    let objectName = "test-csek-resumable"
-    let data = Data(repeating: 1, count: 10 * 1024 * 1024)
-    let source = BytesSource(data: data)
-
-    let sample = sampleKey()
-    let csek = try CustomerEncryptionKeyOptions(key: sample.data)
-
-    let startUrl = registry.url(
-      "/upload/storage/v1/b/\(bucket)/o?uploadType=resumable&name=\(objectName)")
-    let chunkUrl = registry.url("/upload/storage/v1/b/\(bucket)/o?upload_id=csek-upload-id")
-
-    let responseJSON = """
-      {
-        "name": "\(objectName)",
-        "bucket": "\(bucket)",
-        "generation": "1",
-        "metageneration": "1",
-        "size": "\(data.count)",
-        "contentType": "application/octet-stream",
-        "customerEncryption": {
-          "encryptionAlgorithm": "AES256",
-          "keySha256": "\(sample.keyHashBase64)"
-        }
-      }
-      """
-
-    registry.register(
-      response: .success(
-        statusCode: 200, data: Data(),
-        headers: ["Location": chunkUrl.absoluteString]),
-      for: startUrl)
-    registry.register(
-      response: .success(
-        statusCode: 200, data: responseJSON.data(using: .utf8)!,
-        headers: nil),
-      for: chunkUrl)
-
-    let config = URLSessionConfiguration.ephemeral
-    config.protocolClasses = [MockURLProtocol.self]
-    let session = URLSession(configuration: config)
-
-    let options = StorageClientOptions().with {
-      $0.client = .init().with {
-        $0.endpoint = registry.endpoint
-        $0._testSession = session
-      }
-    }
-
-    let client = try StorageClient(options)
-    let uploadOptions = UploadOptions(customerEncryptionKey: csek)
-    let task = client.upload(source, to: bucket, as: objectName, options: uploadOptions)
-    let object = try await task.value
-
-    #expect(object.name == objectName)
-    #expect(object.customerEncryption?.keySha256 == sample.keyHashBase64)
-
-    let requests = registry.recordedRequests()
-    #expect(requests.count == 2)
-
-    // Verify start request headers
-    let startReq = requests[0]
-    #expect(startReq.value(forHTTPHeaderField: "x-goog-encryption-algorithm") == "AES256")
-    #expect(startReq.value(forHTTPHeaderField: "x-goog-encryption-key") == sample.keyBase64)
-    #expect(
-      startReq.value(forHTTPHeaderField: "x-goog-encryption-key-sha256") == sample.keyHashBase64)
-
-    // Verify chunk PUT request headers
-    let chunkReq = requests[1]
-    #expect(chunkReq.value(forHTTPHeaderField: "x-goog-encryption-algorithm") == "AES256")
-    #expect(chunkReq.value(forHTTPHeaderField: "x-goog-encryption-key") == sample.keyBase64)
-    #expect(
-      chunkReq.value(forHTTPHeaderField: "x-goog-encryption-key-sha256") == sample.keyHashBase64)
-  }
 }
